organize payments vertically
