+++ template/app/src/payment/stripe/checkoutUtils.ts
