+++ template/app/src/payment/stripe/stripeClient.ts
